refactor messaging init
