+++ modules/utils/src/index.ts
@@ -11,3 +11,4 @@ export * from "./identifiers";
 export * from "./json";
 export * from "./logger";
 export * from "./math";
+export * from "./url";
